+++ src/obr/OpenFOAM/case.py
@@ -11,7 +11,15 @@ from datetime import datetime
 from Owls.parser.FoamDict import FileParser
 from Owls.parser.LogFile import LogFile
 
-from ..core.core import logged_execute, logged_func, modifies_file, path_to_key
+from ..core.core import (
+    logged_execute,
+    logged_func,
+    modifies_file,
+    path_to_key,
+    TemporaryFolder,
+    DelinkFolder,
+    find_time_folder,
+)
 from .BlockMesh import BlockMesh, calculate_simple_partition
 
 OF_HEADER_REGEX = r"""(/\*--------------------------------\*- C\+\+ -\*----------------------------------\*\\
@@ -163,18 +171,7 @@ class OpenFOAMCase(BlockMesh):
     @property
     def time_folder(self) -> list[Path]:
         """Returns all timestep folder"""
-
-        def is_time(s: str) -> bool:
-            try:
-                float(s)
-                return True
-            except:
-                return False
-
-        _, fs, _ = next(os.walk(self.path))
-        ret = [self.path / f for f in fs if is_time(f)]
-        ret.sort()
-        return ret
+        return find_time_folder(self.path)
 
     @property
     def processor_folder(self) -> list[Path]:
@@ -270,9 +267,35 @@ class OpenFOAMCase(BlockMesh):
     def _exec_operation(self, operation) -> Path:
         return logged_execute(operation, self.path, self.job.doc)
 
+    @property
+    def esi_version(self) -> bool:
+        """Check if esi version of OpenFOAM is sourced"""
+        wm_project_dir = os.environ.get("WM_PROJECT_DIR")
+        if not wm_project_dir:
+            raise AssertionError("OpenFOAM not sourced. Cannot check OpenFOAM version")
+        return (Path(wm_project_dir) / "CONTRIBUTORS.md").exists()
+
     def decomposePar(self, args={}):
         """Sets decomposeParDict and calls decomposePar. If no decomposeParDict exists a new one
         gets created"""
+        tmp_zero = None
+        if not self.time_folder:
+            logging.warning(
+                f"No time folder found! Decomposition might lead to an unusable case."
+            )
+            zero_orig_path = self.path / "0.orig"
+            if zero_orig_path.exists():
+                logging.warning(f"Using existing 0.orig folder")
+                zero_target_path = self.path / "0"
+                tmp_zero = TemporaryFolder(
+                    zero_orig_path, zero_target_path, self.esi_version
+                )
+
+        constant_folder = None
+        if not self.esi_version:
+            constant_folder = DelinkFolder(self.constant_folder)
+            # set a dummy member to avoid issues with autoflake
+            constant_folder.dummy = None
 
         if not self.decomposeParDict:
             decomposeParDictFile = Path(self.system_folder / "decomposeParDict")
@@ -312,6 +335,12 @@ class OpenFOAMCase(BlockMesh):
         fvSolutionArgs = args.get("fvSolution", {})
         if fvSolutionArgs:
             self.fvSolution.set(fvSolutionArgs)
+
+        if not find_time_folder(self.path / "processor0"):
+            logging.warning(
+                f"No time in processor folder found. This indicates an unusable"
+                f" decomposition."
+            )
         return log
 
     def setKeyValuePair(self, args: dict):